Refactor: Introduce RuleTestBase class
The analyzer tests won't be able to use the same test infrastructure as the existing rule tests. We introduce an intermediate base class RuleTestBase from which the existing rule-based tests will continue to derive (until the rules are all gone). In the next commit, we will introduce a parallel base class for the upcoming analyzer-based tests.

The structure will be:

CodeFormatterTestBase
    AnalyzerTestBase
        MyAnalyzerTest
    RuleTestBase
        SyntaxRuleTestBase
            MySyntaxRuleTest
        LocalSemanticRuleTestBase
            MyLocalSemanticRuleTest
        GlobalSemanticRuleTestBase
            MyGlobalSemanticRuleTest