Bring code up to date with Roslyn 1.0.0
The previous commit only changes the packages.config files. It didn't update the project files to actually use the newer versions, and it didn't update the code for the API changes in the new versions. That's now done.

With this commit the unit tests pass, but the command line program itself fails because it's attempting to use reflection to invoke a FixAllContext constructor that no longer exists. I'll fix that separately in the next commit.