Introduce FormattingEngineDiagnosticProvider
This allows us to use a public constructor from the FixAllContext class, instead of using reflection to invoke a private constructor that no longer existed.

Now the command line program gets further. Now it crashes when attempting to apply changes to the workspace, with the message "Changing parse options is not supported".