Get command line program working again
Use the newly available public constructor for FixAllContext. This requires creating a class derived from FixAllContext.DiagnosticProvider. We supply one which incorporates the logic that was previously located in the delegate passed to CreateFixAllContext (which is now gone, replaced by a direct call to the constructor).

To get this to work we had to get rid of the symbol DOTNET_FORMATTER that was being added to the solution. Its presence caused a failure when attempting to apply the changes to the workspace: "Changing parse options is not supported".